docs: add "Run on Apify" for Python example
