+++ website/src/pages/index.js
@@ -13,6 +13,8 @@ import LanguageInfoWidget from '../components/Homepage/LanguageInfoWidget';
 import LanguageSwitch from '../components/Homepage/LanguageSwitch';
 import RunnableCodeBlock from '../components/RunnableCodeBlock';
 
+import PythonHomePageExample from '!!raw-loader!roa-loader!./home_page_example.py';
+
 function LanguageGetStartedSection() {
     return (
         <section className={styles.languageGetStartedSection}>
@@ -67,50 +69,6 @@ const data = await crawler.getData();
 console.table(data.items);
 `;
 
-const pythonExample = `import asyncio
-
-from crawlee.crawlers import PlaywrightCrawler, PlaywrightCrawlingContext
-
-
-async def main() -> None:
-    crawler = PlaywrightCrawler(
-        max_requests_per_crawl=5,  # Limit the crawl to 5 requests at most.
-        headless=False,  # Show the browser window.
-        browser_type='firefox',  # Use the Firefox browser.
-    )
-
-    # Define the default request handler, which will be called for every request.
-    @crawler.router.default_handler
-    async def request_handler(context: PlaywrightCrawlingContext) -> None:
-        context.log.info(f'Processing {context.request.url} ...')
-
-        # Extract and enqueue all links found on the page.
-        await context.enqueue_links()
-
-        # Extract data from the page using Playwright API.
-        data = {
-            'url': context.request.url,
-            'title': await context.page.title(),
-            'content': (await context.page.content())[:100],
-        }
-
-        # Push the extracted data to the default dataset.
-        await context.push_data(data)
-
-    # Run the crawler with the initial list of URLs.
-    await crawler.run(['https://crawlee.dev'])
-
-    # Export the entire dataset to a JSON file.
-    await crawler.export_data('results.json')
-
-    # Or work with the data directly.
-    data = await crawler.get_data()
-    crawler.log.info(f'Extracted data: {data.items}')
-
-
-if __name__ == '__main__':
-    asyncio.run(main())`;
-
 function CodeExampleSection() {
     const [activeOption, setActiveOption] = useState('JavaScript');
     return (
@@ -142,7 +100,9 @@ function CodeExampleSection() {
                         </RunnableCodeBlock>
                     )}
                     {activeOption === 'Python' && (
-                        <CodeBlock language="python">{pythonExample}</CodeBlock>
+                        <RunnableCodeBlock className={styles.codeBlock} type="python" language="python">
+                            {PythonHomePageExample}
+                        </RunnableCodeBlock>
                     )}
                 </div>
                 <div className={styles.dashedSeparator} />
